renames FavoriteShabads to FavouriteShabads
