--- src/js/pages/FavouriteShabads/components/Shabad.tsx
@@ -1,19 +1,0 @@
-/* eslint-disable react/prop-types */
-import React from 'react'
-import { IFavoriteShabad } from '@/types/favorite-shabads';
-
-interface ShabadProps {
-  data: IFavoriteShabad
-}
-
-const Shabad: React.FC<ShabadProps> = ({ data }) => {
-  return (
-    <li className="favourite-shabads--row">
-      <div>{data.id}</div>
-      <div>{data.verse}</div>
-      <div>{data?.source}</div>
-    </li>
-  )
-}
-
-export default Shabad
